+++ docusaurus/docs/recipes/splash-screen.md
@@ -20,7 +20,7 @@ npm i @moxy/react-wait-for-react
 
 ### 2. Add `<SplashScreen />` to your project
 
-Copy the [`splash-screen`](https://github.com/moxystudio/next-with-moxy/tree/master/docusaurus/static/recipes-assets/splash-screen) component folder into `www/shared/components` and edit `www/shared/components/index.js` to export it:
+Copy the [`splash-screen`](https://github.com/moxystudio/next-with-moxy/tree/master/docusaurus/static/recipes-assets/splash-screen) component folder into `www/shared/react` and edit `www/shared/react/index.js` to export it:
 
 ```js
 // ....
@@ -98,7 +98,7 @@ Assuming you integrated the splash screen into the `<MainLayout />` as exemplifi
 ```js
 // ...
 import { withLayout } from '@moxy/next-layout';
-import { MainLayout } from '../../shared/components';
+import { MainLayout } from '../../shared/react';
 
 // ...
