+++ www/pages/home/Home.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import Home from './Home';
-import { AppTree } from '../../shared/test-utils/components';
+import { AppTree } from '../../shared/test-utils/react';
 
 it('should render correctly', () => {
     const { container } = render(<AppTree><Home /></AppTree>);
