+++ www/shared/react/cookie-banner/CookieBanner.js
