+++ www/shared/react/cookie-banner/index.js
