+++ www/shared/react/footer/index.js
