+++ www/shared/react/header/Header.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import Header from './Header';
-import { AppTree } from '../../test-utils/components';
+import { AppTree } from '../../test-utils/react';
 
 it('should render correctly', () => {
     const { container } = render((
